[analyze] Use exceptions to propagate errors during declaration lookup
An identifier cannot be declared more than once, so having a return type
of a sequence of locations just because the protocol supports it doesn't
feel very good. Moving to a single location as the return type requires
us to decide on an error handling model. We use exceptions because using
something like an empty uri to signal an error seems convoluted.